Hold every nutrient figure as a Decimal
A figure on a label is a decimal that was printed, not a measured binary
quantity. Read as a float it scales into noise -- 0.7 g of fat over 40 g came
out 0.27999999999999997 -- which a record and then a recipe file carried
verbatim.

`parse_float=Decimal` keeps the digits a line states, so a read-modify-write
round trip is exact by construction rather than by float formatting being
reproduced. Restating multiplies before dividing, and dividing by a power of
ten becomes a shift, so the only rounding left is the one an inexact division
genuinely needs. The single float boundary is named: `_MAX_FIGURE`, the
largest figure a JSON consumer reading doubles can hold.

+++ src/pantry/commands/lookup.py
@@ -1,5 +1,7 @@
 """`pantry lookup` — exact composite-identity lookup."""
 
+from decimal import Decimal
+
 import click
 from agentcli import json_option
 
@@ -27,7 +29,7 @@ def lookup(
     ctx: click.Context,
     source: str,
     product_id: str,
-    grams: float | None,
+    grams: Decimal | None,
     json_output: bool,
 ) -> None:
     """Find exactly SOURCE and PRODUCT_ID, with no fuzz and no network.
